Add the logic to handle empty unfinished partition when refresh cache
The global watermark should move forward. But we have client report that
the connector terminated due to 0 unfinished partition returned. Adding
this logic to handle the case when the global watermark move backward.
If this never happens, the new logic will not be executed, so it will
not break anything working.